requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
psycopg2==2.9.11
pydantic==2.13.3
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.58.0
shellingham==1.5.4
SQLAlchemy==2.0.49
starlette==1.0.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
